+++ src/Opulence/Framework/Http/Middleware/Session.php
@@ -76,6 +76,7 @@ abstract class Session implements IMiddleware
      * Writes any session data needed in the response
      *
      * @param Response $response The response to write to
+     * @return Response The response with data written to it
      */
     abstract protected function writeToResponse(Response $response);