+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<h2>v1.0.0-alpha29</h2>
+
+<h3>HTTP</h3>
+* Added `HTTP_ACCEPT` default server var to `Request`
+* Changed `Cookie` to accept either a `DateTime` or an int
+
 <h2>v1.0.0-alpha28</h2>
 
 <h3>Testing</h3>